+++ src/main/java/kitchenpos/ordertable/controller/TableRestController.java
@@ -32,16 +32,6 @@ public class TableRestController {
                 ;
     }
 
-    @PutMapping("/api/tables/{orderTableId}/empty")
-    public ResponseEntity<OrderTable> changeEmpty(
-            @PathVariable final Long orderTableId,
-            @RequestBody final OrderTable orderTable
-    ) {
-        return ResponseEntity.ok()
-                .body(tableBo.changeEmpty(orderTableId, orderTable))
-                ;
-    }
-
     @PutMapping("/api/tables/{orderTableId}/number-of-guests")
     public ResponseEntity<OrderTable> changeNumberOfGuests(
             @PathVariable final Long orderTableId,
