+++ src/main/java/kitchenpos/ordertable/domain/Guest.java
@@ -1,0 +1,12 @@
+package kitchenpos.ordertable.domain;
+
+public class Guest {
+    final private int numberOfGuests;
+
+    public Guest(int numberOfGuests) {
+        this.numberOfGuests = numberOfGuests;
+    }
+    public int getNumberOfGuests() {
+        return this.numberOfGuests;
+    }
+}
