+++ src/test/java/kitchenpos/bo/OrderBoTest.java
@@ -3,15 +3,19 @@ package kitchenpos.bo;
 import kitchenpos.menu.domain.MenuDao;
 import kitchenpos.order.domain.OrderDao;
 import kitchenpos.order.domain.OrderLineItemDao;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.bo.OrderBo;
+import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.ordertable.domain.TableGroupDao;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.Arrays;
 import java.util.List;
@@ -26,6 +30,8 @@ class OrderBoTest {
     private final OrderDao orderDao = new InMemoryOrderDao();
     private final OrderLineItemDao orderLineItemDao = new InMemoryOrderLineItemDao();
     private final OrderTableDao orderTableDao = new InMemoryOrderTableDao();
+    private final TableGroupDao tableGroupDao = new InMemoryTableGroupDao();
+
 
     private OrderBo orderBo;
 
@@ -124,4 +130,56 @@ class OrderBoTest {
                         .containsExactlyInAnyOrderElementsOf(expected.getOrderLineItems())
         );
     }
+
+    @DisplayName("빈 테이블 설정 또는 해지할 수 있다.")
+    @ParameterizedTest
+    @ValueSource(booleans = {true, false})
+    void changeEmpty(final boolean empty) {
+        // given
+        final Long orderTableId = orderTableDao.save(table1()).getId();
+
+        final OrderTable expected = new OrderTable();
+        expected.setEmpty(empty);
+
+        // when
+        final OrderTable actual = orderBo.changeEmpty(orderTableId, expected);
+
+        // then
+        assertThat(actual).isNotNull();
+        assertAll(
+                () -> assertThat(actual.getId()).isEqualTo(orderTableId),
+                () -> assertThat(actual.isEmpty()).isEqualTo(expected.isEmpty())
+        );
+    }
+
+    @DisplayName("단체 지정을 해지할 수 있다.")
+    @Test
+    void ungroup() {
+        // given
+        final Long tableGroupId = saveTable1AndTable2().getId();
+
+        // when
+        // then
+        orderBo.ungroup(tableGroupId);
+    }
+
+    @DisplayName("단체 지정된 테이블의 주문 상태가 조리 또는 식사인 경우 단체 지정을 해지할 수 없다.")
+    @Test
+    void ungroupNotCalculatedTableGroup() {
+        // given
+        final Long tableGroupId = saveTable1AndTable2().getId();
+        orderDao.save(orderForTable1());
+
+        // when
+        // then
+        assertThatExceptionOfType(IllegalArgumentException.class).isThrownBy(() -> orderBo.ungroup(tableGroupId));
+    }
+
+    private TableGroup saveTable1AndTable2() {
+        final TableGroup tableGroup = tableGroupDao.save(table1AndTable2());
+        orderTableDao.save(groupedTable1());
+        orderTableDao.save(groupedTable2());
+        return tableGroup;
+    }
+
 }
