+++ src/test/java/kitchenpos/bo/TableGroupBoTest.java
@@ -15,7 +15,6 @@ import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 class TableGroupBoTest {
-    private final OrderDao orderDao = new InMemoryOrderDao();
     private final OrderTableDao orderTableDao = new InMemoryOrderTableDao();
     private final TableGroupDao tableGroupDao = new InMemoryTableGroupDao();
 
@@ -23,7 +22,7 @@ class TableGroupBoTest {
 
     @BeforeEach
     void setUp() {
-        tableGroupBo = new TableGroupBo(orderDao, orderTableDao, tableGroupDao);
+        tableGroupBo = new TableGroupBo(orderTableDao, tableGroupDao);
         orderTableDao.save(emptyTable1());
         orderTableDao.save(emptyTable2());
     }
@@ -58,33 +57,4 @@ class TableGroupBoTest {
         assertThatExceptionOfType(IllegalArgumentException.class).isThrownBy(() -> tableGroupBo.create(expected));
     }
 
-    @DisplayName("단체 지정을 해지할 수 있다.")
-    @Test
-    void ungroup() {
-        // given
-        final Long tableGroupId = saveTable1AndTable2().getId();
-
-        // when
-        // then
-        tableGroupBo.ungroup(tableGroupId);
-    }
-
-    @DisplayName("단체 지정된 테이블의 주문 상태가 조리 또는 식사인 경우 단체 지정을 해지할 수 없다.")
-    @Test
-    void ungroupNotCalculatedTableGroup() {
-        // given
-        final Long tableGroupId = saveTable1AndTable2().getId();
-        orderDao.save(orderForTable1());
-
-        // when
-        // then
-        assertThatExceptionOfType(IllegalArgumentException.class).isThrownBy(() -> tableGroupBo.ungroup(tableGroupId));
-    }
-
-    private TableGroup saveTable1AndTable2() {
-        final TableGroup tableGroup = tableGroupDao.save(table1AndTable2());
-        orderTableDao.save(groupedTable1());
-        orderTableDao.save(groupedTable2());
-        return tableGroup;
-    }
 }
